requirements.txt
django==3.0.7
djangorestframework==3.11.0
markdown==3.2.2
factory-boy==2.12.0
